Make ArrivalQueryable independent of GTFS provider
